Rename: 스펠링 오타 수정 (schdule -> schedule)
