+++ src/entities/group/api/get-group-details.ts
@@ -1,9 +1,9 @@
-// src/entities/group/api/get-group-details.ts
+import { get } from "@/features/auth/login";
 import { BASE_URL } from "@/shared/constants";
 import { GroupDetails } from "../model/types";
 
 export const getGroupDetails = async (groupId: number): Promise<GroupDetails> => {
-    const response = await fetch(`${BASE_URL}/api/v1/groups/${groupId}`);
+    const response = await get(`${BASE_URL}/api/v1/groups/${groupId}`);
     const data = await response.json();
 
     if (data.code !== 200) {
