+++ src/entities/schedule/api/get-my-schedules.ts
