+++ src/entities/schedule/api/get-schedule-details.ts
