+++ src/entities/schedule/api/schedule.query.ts
