+++ src/entities/schedule/model/types.ts
