+++ src/entities/schedule/ui/my-schedule-card.tsx
