+++ src/entities/schedule/ui/schedule-card.tsx
