--- src/widgets/schdule-list/index.ts
@@ -1,5 +1,0 @@
-// ui
-export * from './ui/schedule-list'
-
-// api
-export * from '../../entities/schdule/api/get-schedules'
