+++ src/widgets/schedule-list/index.ts
@@ -1,0 +1,5 @@
+// ui
+export * from './ui/schedule-list'
+
+// api
+export * from '../../entities/schedule/api/get-schedules'
