+++ src/widgets/schedule-list/ui/schedule-list.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { ScheduleCard, scheduleQuery } from "@/entities/schdule";
+import { ScheduleCard, scheduleQuery } from "@/entities/schedule";
 import { useInfiniteQuery } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
 import { useMemo } from "react";
